Changed Menu Controller for some fixes. It would've beeped every time something was pressed, but now only when enabled.
Added Main Menu Controller. Tested, Fully Functional.

Changed Scene Manager to load based on the current scene's name. We change this definition in the future to load future scenes.

Will properly load/save/change profile data. We will use SaveData0 as the temporary SaveData. When Loading Game, we will save the session Number to the profile Number. The New Scene will indirectly accesses this profile Number to load the profile.

Changes to volume and resolution depend on if the settings were changed in the Main Menu.

When built, will now set Main Menu Scene as the first one. We can then load Test Scene and check if the profile was properly loaded or if the settings were properly changed.